Moved datasets to their own sub-package. Added CMU ARCTIC dataset support.
