+++ pyroomacoustics/datasets/__init__.py
@@ -1,0 +1,3 @@
+
+from .timit import Word, Sentence, TimitCorpus
+from .cmu_arctic import CMUArcticCorpus, CMUArcticSentence
